fix: delete an extra presenter
