+++ src/presenter/point-presenter.js
@@ -2,7 +2,6 @@ import TripEventsListView from '../view/trip-events-list-view.js';
 import RoutPointView from '../view/route-point-view.js';
 import {remove, render, replace} from '../framework/render.js';
 import RoutPointEditView from '../view/route-point-edit-view.js';
-import RoutPointBoxView from '../view/route-point-box-view.js';
 
 
 const Mode = {
@@ -13,7 +12,6 @@ const Mode = {
 export default class PointPresenter {
   #listComponent = new TripEventsListView;
   #taskEditComponent = null;
-  #taskBoxComponent = null;
   #taskComponent = null;
   #task = null;
   #handleDataChange = null;
@@ -29,9 +27,9 @@ export default class PointPresenter {
   init(task) {
     this.#task = task;
 
-    const prevTaskComponent = this.#taskComponent;
-    const prevTaskEditComponent = this.#taskEditComponent;
-    const prevTaskBoxComponent = this.#taskBoxComponent;
+    // const prevTaskComponent = this.#taskComponent;
+    // const prevTaskEditComponent = this.#taskEditComponent;
+    // const prevTaskBoxComponent = this.#taskBoxComponent;
 
     this.#taskEditComponent = new RoutPointEditView({
       task: this.#task,
@@ -41,7 +39,7 @@ export default class PointPresenter {
       }
     });
 
-    this.#taskBoxComponent = new RoutPointBoxView({
+    this.#taskComponent = new RoutPointView({
       task: this.#task,
       onEditClick: () => {
         this.#replaceCardToForm();
@@ -51,42 +49,27 @@ export default class PointPresenter {
       },
     });
 
-    this.#taskComponent = new RoutPointView({
-      task: this.#task,
-      taskEditComponent: this.#taskEditComponent,
-      taskBoxComponent: this.#taskBoxComponent,
-    });
+    // if (prevTaskComponent === null || prevTaskEditComponent === null || prevTaskBoxComponent === null) {
+    render(this.#taskComponent, this.#listComponent);
+    //   return;
+    // }
 
-    if (prevTaskComponent === null || prevTaskEditComponent === null || prevTaskBoxComponent === null) {
-      render(this.#taskComponent, this.#listComponent);
-      render(this.#taskBoxComponent, this.#taskComponent.element);
-      return;
-    }
+    // if (this.#listComponent.contains(prevTaskComponent.element)) {
+    //   replace(this.#taskComponent, prevTaskComponent);
+    // }
 
-    if (this.#listComponent.contains(prevTaskComponent.element)) {
-      // console.log(prevTaskComponent.element);
-      replace(this.#taskComponent, prevTaskComponent);
-    }
+    // if (this.#mode === Mode.DEFAULT) {
+    //   replace(this.#taskBoxComponent, prevTaskBoxComponent);
+    // }
 
-    if (this.#mode === Mode.DEFAULT) {
-      // console.log(this.#taskBoxComponent);
-      replace(this.#taskBoxComponent, prevTaskBoxComponent);
-    }
-
-    if (this.#mode === Mode.EDITING) {
-      // console.log(this.#taskEditComponent);
-      replace(this.#taskEditComponent, prevTaskEditComponent);
-    }
-
-    remove(prevTaskComponent);
-    remove(prevTaskEditComponent);
-    remove(prevTaskBoxComponent);
+    // if (this.#mode === Mode.EDITING) {
+    //   replace(this.#taskEditComponent, prevTaskEditComponent);
+    // }
   }
 
   destroy() {
     remove(this.#taskComponent);
     remove(this.#taskEditComponent);
-    remove(this.#taskBoxComponent);
   }
 
   resetView() {
@@ -96,14 +79,14 @@ export default class PointPresenter {
   }
 
   #replaceCardToForm() {
-    replace(this.#taskEditComponent, this.#taskBoxComponent);
+    replace(this.#taskEditComponent, this.#taskComponent);
     document.addEventListener('keydown', this.#escKeyDownHandler);
     this.#handleModeChange();
     this.#mode = Mode.EDITING;
   }
 
   #replaceFormToCard() {
-    replace(this.#taskBoxComponent, this.#taskEditComponent);
+    replace(this.#taskComponent, this.#taskEditComponent);
     document.removeEventListener('keydown', this.#escKeyDownHandler);
     this.#mode = Mode.DEFAULT;
   }
@@ -121,7 +104,6 @@ export default class PointPresenter {
   };
 
   #handleFavoriteClick = () => {
-    // console.log({...this.#task, isFavorite: !this.#task.isFavorite});
     this.#handleDataChange({...this.#task, isFavorite: !this.#task.isFavorite});
   };
